chore: resolve the accumulated SonarQube findings
Sweep open issues down to the deliberately kept ones.

Deliberate patterns the analyzer cannot judge keep their behavior
and get a targeted suppression with a one-line rationale.

on-behalf-of: @multiversio <info@multivers.io>

+++ parquetry-compression/src/main/java/io/tileverse/parquetry/compression/Codec.java
@@ -360,7 +360,7 @@ public sealed interface Codec
                         target = grown;
                     }
                     ByteBuffer window =
-                            target.asSlice(produced, capacity - produced).asByteBuffer();
+                            target.asSlice(produced, (long) capacity - produced).asByteBuffer();
                     int n = inflater.inflate(window);
                     if (n == 0) {
                         if (inflater.finished()) {
